pyproject.toml
[project]
name = "meraki-usecase"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "urllib3<2",
  "meraki",
  "python-dotenv>=1.0.0",
]

[project.scripts]
meraki-usecase = "meraki_usecase.cli:main"
